integrated user into watchlist

+++ controllers/watchlist.php
@@ -1,5 +1,7 @@
 <?php
 
+use GW2Spidy\DB\User;
+
 use \DateTime;
 
 use GW2Spidy\Application;
@@ -33,19 +35,27 @@ use GW2Spidy\Util\Functions;
  *  route /watchlistadd POST
  * ----------------------
  */
-$app->post("/watchlistadd", function (Request $request) use ($app) {    
+$app->post("/watchlist/add", function (Request $request) use ($app) {
+    if (!$app['security']->getToken() || !($user = $app['security']->getToken()->getUser()) || !$user instanceof User) {
+        return $app->redirect($app['url_generator']->generate('login'));
+    }
+
     $data_id = $request->get("data_id");
     if($data_id){
-        
+
         // TODO: Test if item is valid item!
-        
-        $w = new Watchlist();
-        $w->setUserId(1); //TODO: replace with real user
-        $w->setItemId($data_id);
-        $w->save();
+
+        // check unique
+        if (!WatchlistQuery::create()->filterByUser($user)->filterByItemId($data_id)->count()) {
+            $w = new Watchlist();
+            $w->setUser($user);
+            $w->setItemId($data_id);
+
+            $w->save();
+        }
     }
     //TODO: check what happens when referer is disabled in browser
-    $uri = $request->headers->get('referer');
+    $uri = $request->headers->get('referer') ?: $app['url_generator']->generate('item', array('dataId' => $data_id));
     return $app->redirect($uri);
 })
 ->bind('watchlistaddpost');
@@ -55,15 +65,19 @@ $app->post("/watchlistadd", function (Request $request) use ($app) {
  *  route /watchlistremove POST
  * ----------------------
  */
-$app->post("/watchlistremove", function (Request $request) use ($app) {    
+$app->post("/watchlist/remove", function (Request $request) use ($app) {
+    if (!$app['security']->getToken() || !($user = $app['security']->getToken()->getUser()) || !$user instanceof User) {
+        return $app->redirect($app['url_generator']->generate('login'));
+    }
+
     $data_id = $request->get("data_id");
     if($data_id){
-        
+
         // TODO: Test if item is valid item!
-        
+
         $w = WatchlistQuery::create();
         $w->filterByItemId($data_id);
-        $w->filterByUserId(1);  //TODO: replace with real user
+        $w->filterByUser($user);
         $w->delete();
     }
     return $app->redirect($app['url_generator']->generate('watchlist'));
@@ -76,20 +90,26 @@ $app->post("/watchlistremove", function (Request $request) use ($app) {
  * ----------------------
  */
 $app->get("/watchlist/{page}", function(Request $request, $page) use($app) {
+    if (!$app['security']->getToken() || !($user = $app['security']->getToken()->getUser()) || !$user instanceof User) {
+        return $app->redirect($app['url_generator']->generate('login'));
+    }
+
     $itemIds = array();
-    $w = WatchlistQuery::create()->findByUserId(1);
-    if($w){
-        $itemIds = array();
-        foreach($w as $item){
-            $itemIds[] = $item->getItemId();
+    if ($ws = $user->getWatchlists()) {
+        foreach ($ws as $w) {
+            $itemIds[] = $w->getItemId();
         }
     }
+
     $page = $page > 0 ? $page : 1;
+
     $q = ItemQuery::create();
-    $q->filterByPrimaryKeys($itemIds);
+    $q->filterByPrimaryKeys(array_unique($itemIds));
+
     // use generic function to render
-    return item_list($app, $request, $q, $page, 25,array('watchlist'=> true));
+    return item_list($app, $request, $q, $page, 25, array('watchlist'=> true));
 })
-->value('page',      1)
+->assert('page', '\d*')
+->value('page', 1)
 ->bind('watchlist');
